# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Central authentication microservice — SSO for all abuamar.online projects.",
        "title": "Auth Service API",
        "termsOfService": "https://abuamar.online/terms",
        "contact": {
            "name": "Abu Amar",
            "email": "abuamar.albadawi@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "auth.abuamar.online",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticate with email/username + password. Returns JWT access + refresh tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoke a refresh token (invalidates the session).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "description": "Logout payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's profile.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerUserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Exchange a refresh token for a new access + refresh token pair (rotation).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh tokens",
                "parameters": [
                    {
                        "description": "Refresh payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Create a new user account. At least one of email or username is required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_abuamar142_auth-service_internal_models.LoginRequest": {
            "type": "object",
            "properties": {
                "identifier": {
                    "type": "string",
                    "example": "test@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "secretpass123"
                }
            }
        },
        "github_com_abuamar142_auth-service_internal_models.LogoutRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "e1d2a3b4c5f6a7b8c9d0e1f2a3b4c5d6a7b8c9d0e1f2a3b4c5d6a7b8c9d0e1f2"
                }
            }
        },
        "github_com_abuamar142_auth-service_internal_models.RefreshRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "e1d2a3b4c5f6a7b8c9d0e1f2a3b4c5d6a7b8c9d0e1f2a3b4c5d6a7b8c9d0e1f2"
                }
            }
        },
        "github_com_abuamar142_auth-service_internal_models.RegisterRequest": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "type": "string",
                    "example": "test@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "secretpass123"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
        "github_com_abuamar142_auth-service_internal_models.SwaggerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "string",
                    "example": "field is required"
                }
            }
        },
        "github_com_abuamar142_auth-service_internal_models.SwaggerResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerError"
                },
                "message": {
                    "type": "string",
                    "example": "operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_abuamar142_auth-service_internal_models.SwaggerTokenResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "access_token": {
                            "type": "string",
                            "example": "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyZDk0NGVkNi0yODVlLTRjMTMtODkxMS04OTUyMGMzNGU0M2QiLCJlbWFpbCI6InRlc3RAZXhhbXBsZS5jb20iLCJ1c2VybmFtZSI6ImpvaG5kb2UiLCJpYXQiOjE3ODk1NDcwNTksImV4cCI6MTc4OTU0Nzk1OX0.RWWXCR49EGPyZvVeqDTmKAdn3pnpXXWK1yx9Jk5-BRY"
                        },
                        "refresh_token": {
                            "type": "string",
                            "example": "54204a5db8e34b88874032e4b63add0e538747c0e498f4db7217b706dda18c24"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "login successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_abuamar142_auth-service_internal_models.SwaggerUser": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-09-16T08:23:55.441437Z"
                },
                "display_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "type": "string",
                    "example": "test@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "2d944ed6-285e-4c13-8911-89520c34e43d"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-09-16T08:23:55.441437Z"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "github_com_abuamar142_auth-service_internal_models.SwaggerUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_abuamar142_auth-service_internal_models.SwaggerUser"
                },
                "message": {
                    "type": "string",
                    "example": "user registered"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "Enter your API key",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}