# internal/handlers/auth.go
package handlers

import (
	"encoding/json"
	"errors"
	"net/http"
	"strings"
	"time"

	"github.com/abuamar142/auth-service/internal/middleware"
	_ "github.com/abuamar142/auth-service/internal/models"
	"github.com/abuamar142/auth-service/internal/response"
	"github.com/abuamar142/auth-service/internal/services"
)

type AuthHandler struct {
	AuthService *services.AuthService
}

func NewAuthHandler(svc *services.AuthService) *AuthHandler {
	return &AuthHandler{AuthService: svc}
}

// Register godoc
// @Summary      Register a new user
// @Description  Create a new user account. At least one of email or username is required.
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        body body models.RegisterRequest true "Registration payload"
// @Success      201 {object} models.SwaggerUserResponse
// @Failure      400 {object} models.SwaggerResponse
// @Failure      409 {object} models.SwaggerResponse
// @Router       /api/v1/auth/register [post]
func (h *AuthHandler) Register(w http.ResponseWriter, r *http.Request) {
	var req struct {
		Email       string `json:"email"`
		Username    string `json:"username"`
		Password    string `json:"password"`
		DisplayName string `json:"display_name"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		response.Error(w, http.StatusBadRequest, "INVALID_JSON", "invalid request body", "")
		return
	}
	if req.Password == "" {
		response.Error(w, http.StatusBadRequest, "VALIDATION_ERROR", "password is required", "")
		return
	}
	if len(req.Password) < 8 {
		response.Error(w, http.StatusBadRequest, "VALIDATION_ERROR", "password too short", "must be at least 8 characters")
		return
	}

	user, err := h.AuthService.Register(r.Context(), req.Email, req.Username, req.Password, req.DisplayName)
	if err != nil {
		switch {
		case errors.Is(err, services.ErrIdentifierRequired):
			response.Error(w, http.StatusBadRequest, "VALIDATION_ERROR", err.Error(), "")
		case errors.Is(err, services.ErrEmailExists), errors.Is(err, services.ErrUsernameExists):
			response.Error(w, http.StatusConflict, "IDENTIFIER_TAKEN", "email or username already taken", "")
		default:
			response.Error(w, http.StatusInternalServerError, "INTERNAL_ERROR", "failed to register user", "")
		}
		return
	}

	response.JSON(w, http.StatusCreated, "user registered", user)
}

// Login godoc
// @Summary      Login
// @Description  Authenticate with email/username + password. Returns JWT access + refresh tokens.
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        body body models.LoginRequest true "Login payload"
// @Success      200 {object} models.SwaggerTokenResponse
// @Failure      400 {object} models.SwaggerResponse
// @Failure      401 {object} models.SwaggerResponse
// @Router       /api/v1/auth/login [post]
func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {
	var req struct {
		Identifier string `json:"identifier"`
		Password   string `json:"password"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		response.Error(w, http.StatusBadRequest, "INVALID_JSON", "invalid request body", "")
		return
	}
	if req.Identifier == "" || req.Password == "" {
		response.Error(w, http.StatusBadRequest, "VALIDATION_ERROR", "identifier and password are required", "")
		return
	}

	accessToken, refreshToken, err := h.AuthService.Login(r.Context(), req.Identifier, req.Password)
	if err != nil {
		switch {
		case errors.Is(err, services.ErrInvalidCredentials):
			response.Error(w, http.StatusUnauthorized, "INVALID_CREDENTIALS", "invalid email/username or password", "")
		default:
			response.Error(w, http.StatusInternalServerError, "INTERNAL_ERROR", "failed to login", "")
		}
		return
	}

	response.JSON(w, http.StatusOK, "login successful", map[string]string{
		"access_token":  accessToken,
		"refresh_token": refreshToken,
	})
}

// Refresh godoc
// @Summary      Refresh tokens
// @Description  Exchange a refresh token for a new access + refresh token pair (rotation).
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        body body models.RefreshRequest true "Refresh payload"
// @Success      200 {object} models.SwaggerTokenResponse
// @Failure      400 {object} models.SwaggerResponse
// @Failure      401 {object} models.SwaggerResponse
// @Router       /api/v1/auth/refresh [post]
func (h *AuthHandler) Refresh(w http.ResponseWriter, r *http.Request) {
	var req struct {
		RefreshToken string `json:"refresh_token"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		response.Error(w, http.StatusBadRequest, "INVALID_JSON", "invalid request body", "")
		return
	}
	if req.RefreshToken == "" {
		response.Error(w, http.StatusBadRequest, "VALIDATION_ERROR", "refresh_token is required", "")
		return
	}

	accessToken, refreshToken, err := h.AuthService.Refresh(r.Context(), req.RefreshToken)
	if err != nil {
		switch {
		case errors.Is(err, services.ErrInvalidRefreshToken):
			response.Error(w, http.StatusUnauthorized, "INVALID_REFRESH_TOKEN", "invalid or expired refresh token", "")
		default:
			response.Error(w, http.StatusInternalServerError, "INTERNAL_ERROR", "failed to refresh token", "")
		}
		return
	}

	response.JSON(w, http.StatusOK, "token refreshed", map[string]string{
		"access_token":  accessToken,
		"refresh_token": refreshToken,
	})
}

// Logout godoc
// @Summary      Logout
// @Description  Revoke a refresh token (invalidates the session).
// @Tags         auth
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        body body models.LogoutRequest true "Logout payload"
// @Success      200 {object} models.SwaggerResponse
// @Failure      400 {object} models.SwaggerResponse
// @Failure      401 {object} models.SwaggerResponse
// @Router       /api/v1/auth/logout [post]
func (h *AuthHandler) Logout(w http.ResponseWriter, r *http.Request) {
	user := middleware.GetUser(r.Context())
	if user == nil {
		response.Error(w, http.StatusUnauthorized, "MISSING_TOKEN", "authentication required", "")
		return
	}

	var req struct {
		RefreshToken string `json:"refresh_token"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		response.Error(w, http.StatusBadRequest, "INVALID_JSON", "invalid request body", "")
		return
	}
	if req.RefreshToken == "" {
		response.Error(w, http.StatusBadRequest, "VALIDATION_ERROR", "refresh_token is required", "")
		return
	}

	err := h.AuthService.Logout(r.Context(), req.RefreshToken)
	if err != nil {
		response.Error(w, http.StatusUnauthorized, "INVALID_REFRESH_TOKEN", "invalid or expired refresh token", "")
		return
	}

	response.JSON(w, http.StatusOK, "logged out", nil)
}

// Me godoc
// @Summary      Get current user
// @Description  Returns the authenticated user's profile.
// @Tags         auth
// @Produce      json
// @Security     BearerAuth
// @Success      200 {object} models.SwaggerUserResponse
// @Failure      401 {object} models.SwaggerResponse
// @Router       /api/v1/auth/me [get]
func (h *AuthHandler) Me(w http.ResponseWriter, r *http.Request) {
	user := middleware.GetUser(r.Context())
	if user == nil {
		response.Error(w, http.StatusUnauthorized, "MISSING_TOKEN", "authentication required", "")
		return
	}
	response.JSON(w, http.StatusOK, "user retrieved", user)
}

// parseDuration parses a simple duration string like "24h", "7d", "30d".
func parseDuration(s string) (time.Duration, error) {
	s = strings.TrimSpace(s)
	if s == "" {
		return 0, nil
	}
	if strings.HasSuffix(s, "d") {
		n := 0
		for _, c := range s[:len(s)-1] {
			if c >= '0' && c <= '9' {
				n = n*10 + int(c-'0')
			}
		}
		if n > 0 {
			return time.Duration(n) * 24 * time.Hour, nil
		}
	}
	return time.ParseDuration(s)
}